refactor(gfn): split cloudmatch.ts into focused modules behind facade
Extract transport, signaling, session parsing, session request builders,
and streaming-feature helpers into dedicated modules while preserving the
public cloudmatch API for callers and tests.
